Next

+++ src/openhab_ui.hpp
@@ -7,7 +7,7 @@
 
 void openhab_ui_setup(Config *config);
 void openhab_ui_set_wifi_state(bool wifi_state);
-void openhab_ui_connect(const char *host, int port, const char *sitemap);
+void openhab_ui_connect(const char *host, uint16_t port, const char *sitemap);
 void openhab_ui_loop(void);
 
 #endif
